Use nonnegative-integer? in favor of natural-number?
The natural-number? predicate seems to be
unavailable (though there is a remnant of it in the form of
natural-number/c, which is just a contract alias of
exact-nonnegative-integer?). In any case, it surely means
nonnegative integer.

(One might push the envelope here slightly and add
exactness, but this commit does not go that far.)